moved fetch command to separate folder
